requirements.txt
kafka-python
pyarrow
pyiceberg[glue]

